nl80211: consider p2p when changing vif type
Commit 9f51b11 ("nl80211: use P2P interface types") added
support for p2p interfaces when adding a new interface.
however, it didn't handle the case in which the same
interface is being used and its type is being changed.
Add support for this case.

Consequently, when doing "ap_scan_as_station" we now need
to save the actual ap interface type (normal/p2p) in order
to restore it properly. for that, change ap_scan_as_station
type from int to nl80211_iftype, and set it to
NL80211_IFTYPE_UNSPECIFED when not used.

Signed-off-by: Eliad Peller <eliad@wizery.com>