driver_nl80211: implement hapd_set_ssid callback
When operating in AP mode, this callback passes the SSID to the kernel
driver.

wl12xx cards handle probe requests in FW. When the SSID is not
broadcast in beacon data (i.e. the SSID is hidden), the AP must know the
real SSID in order to respond to directed probe requests.