AP: Add station before sending the association response
There is a race condition between adding an associated sta (which
happens on assoc_rsp tx completion) and getting the first
packet.

In the case of IEEE 802.1X we might drop the first EAPOL Start
packet which will result in unsuccessful connection.

Solve it by adding the associated station before sending the
association response.