wpa_s: check if sched_scan is supported before using it
Previously, we were trying to use scheduled scan without checking
whether the driver supports it.  This is okay, but it is not efficient
because we need to go all the way down to the kernel driver before we
find out it is not supported.

This patch adds a capability flag in wpa_supplicant that is used to
check whether sched_scan is supported or not.

Signed-off-by: Luciano Coelho <coelho@ti.com>