sched scan: Fix passive scanning
Scan wasn't initiated in case the config contained only networks without scan_ssid.
In such a case we want scan to be initiated without any SSIDs to actively scan
but include all the SSIDs in the filter list.
Also added some debug logs to easily see which SSIDs were included in which list.